+++ appBackend/common-service-data/src/jsMain/kotlin/com/gchristov/thecodinglove/commonservicedata/api/ApiRequest.kt
@@ -1,4 +1,4 @@
-package com.gchristov.thecodinglove.commonservicedata
+package com.gchristov.thecodinglove.commonservicedata.api
 
 import kotlinx.serialization.decodeFromString
 import kotlinx.serialization.json.Json
