+++ appBackend/common-service-data/src/jsMain/kotlin/com/gchristov/thecodinglove/commonservicedata/api/JavascriptFirebaseFunctionsExternals.kt
@@ -1,0 +1,12 @@
+package com.gchristov.thecodinglove.commonservicedata.api
+
+import com.gchristov.thecodinglove.commonservicedata.pubsub.FirebaseFunctionsPubSub
+
+external var exports: dynamic
+
+@JsModule("firebase-functions")
+@JsNonModule
+external object FirebaseFunctions {
+    val https: FirebaseFunctionsHttps
+    val pubsub: FirebaseFunctionsPubSub
+}
