+++ appBackend/common-service-data/src/jsMain/kotlin/com/gchristov/thecodinglove/commonservicedata/pubsub/JavascriptGooglePubSubExternals.kt
@@ -1,7 +1,7 @@
 // This is needed, otherwise we get "PubSub is not a constructor" error
 @file:JsModule("@google-cloud/pubsub")
 @file:JsNonModule
-package com.gchristov.thecodinglove.commonservicedata
+package com.gchristov.thecodinglove.commonservicedata.pubsub
 
 import com.gchristov.thecodinglove.kmpcommonkotlin.Buffer
