+++ appBackend/search/src/jsMain/kotlin/com/gchristov/thecodinglove/search/SearchModule.kt
@@ -1,6 +1,6 @@
 package com.gchristov.thecodinglove.search
 
-import com.gchristov.thecodinglove.commonservicedata.PubSub
+import com.gchristov.thecodinglove.commonservicedata.pubsub.PubSub
 import com.gchristov.thecodinglove.kmpcommondi.DiModule
 import com.gchristov.thecodinglove.searchdata.usecase.PreloadSearchResultUseCase
 import com.gchristov.thecodinglove.searchdata.usecase.SearchWithSessionUseCase
