+++ appBackend/slack/src/jsMain/kotlin/com/gchristov/thecodinglove/slack/usecase/VerifySlackRequestUseCase.kt
@@ -1,7 +1,7 @@
 package com.gchristov.thecodinglove.slack.usecase
 
 import arrow.core.Either
-import com.gchristov.thecodinglove.commonservicedata.ApiRequest
+import com.gchristov.thecodinglove.commonservicedata.api.ApiRequest
 
 interface VerifySlackRequestUseCase {
     suspend operator fun invoke(request: ApiRequest): Either<Throwable, Unit>
